sb-client/requirements.txt
PyQt5==5.15.9
psutil==5.9.5
requests==2.31.0
python-dotenv==1.0.0
pyinstaller==6.3.0
Pillow==10.2.0
winshell==0.6
pywin32==306 